+++ sympy/codegen/cutils.py
@@ -1,4 +1,4 @@
-from sympy.printing.ccode import C99CodePrinter
+from sympy.printing.c import C99CodePrinter
 
 def render_as_source_file(content, Printer=C99CodePrinter, settings=None):
     """ Renders a C source file (with required #include statements) """
