+++ sympy/codegen/tests/test_applications.py
@@ -3,7 +3,7 @@
 import tempfile
 
 from sympy.external import import_module
-from sympy.printing.ccode import ccode
+from sympy.printing import ccode
 from sympy.utilities._compilation import compile_link_import_strings, has_c
 from sympy.utilities._compilation.util import may_xfail
 from sympy.testing.pytest import skip
