+++ sympy/codegen/tests/test_fnodes.py
@@ -13,7 +13,7 @@ from sympy.codegen.fnodes import (
 from sympy.codegen.futils import render_as_module
 from sympy.core.expr import unchanged
 from sympy.external import import_module
-from sympy.printing.fcode import fcode
+from sympy.printing import fcode
 from sympy.utilities._compilation import has_fortran, compile_run_strings, compile_link_import_strings
 from sympy.utilities._compilation.util import may_xfail
 from sympy.testing.pytest import skip
