+++ sympy/external/tests/test_autowrap.py
@@ -256,7 +256,7 @@ def test_autowrap_custom_printer():
 
     from sympy import pi
     from sympy.utilities.codegen import C99CodeGen
-    from sympy.printing.ccode import C99CodePrinter
+    from sympy.printing.c import C99CodePrinter
 
     class PiPrinter(C99CodePrinter):
         def _print_Pi(self, expr):
