+++ sympy/printing/tests/test_cxx.py
@@ -1,6 +1,6 @@
 from sympy import symbols
 from sympy.functions import beta, Ei, zeta, Max, Min, sqrt
-from sympy.printing.cxxcode import CXX98CodePrinter, CXX11CodePrinter, CXX17CodePrinter, cxxcode
+from sympy.printing.cxx import CXX98CodePrinter, CXX11CodePrinter, CXX17CodePrinter, cxxcode
 from sympy.codegen.cfunctions import log1p
 
 x, y = symbols('x y')
