+++ sympy/printing/tests/test_fortran.py
@@ -11,7 +11,7 @@ from sympy.codegen.ast import Declaration, Variable, float32, float64, \
 from sympy.core.relational import Relational
 from sympy.logic.boolalg import And, Or, Not, Equivalent, Xor
 from sympy.matrices import Matrix, MatrixSymbol
-from sympy.printing.fcode import fcode, FCodePrinter
+from sympy.printing.fortran import fcode, FCodePrinter
 from sympy.tensor import IndexedBase, Idx
 from sympy.utilities.lambdify import implemented_function
 from sympy.testing.pytest import raises
